requirements.txt
# Core Framework
langgraph==0.2.28
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5

# Web Framework
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# PDF & Image Processing (FREE!)
reportlab==4.2.5
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==11.0.0
poppler-utils==0.1.0

# OCR (FREE!)
pytesseract==0.3.13

# NLP & Text Processing (FREE!)
python-dateutil==2.9.0

# Utilities
python-multipart==0.0.20
python-dotenv==1.0.1
requests==2.32.3
coloredlogs
python-dotenv
